# schemas/sut-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "E-commerce SUT API",
    "version": "1.0.0",
    "description": "Business contract every architecture profile (A01-A12) must implement identically. Monolith/REST checkout returns 201; event-driven returns 202 and completes asynchronously."
  },
  "components": {
    "parameters": {
      "RequestId": {
        "name": "X-Request-Id",
        "in": "header",
        "required": true,
        "schema": { "type": "string", "pattern": "^[A-Za-z0-9._:-]{1,128}$" }
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 80,
          "pattern": "^[A-Za-z0-9._:-]+$"
        }
      },
      "UserIdHeader": {
        "name": "X-User-Id",
        "in": "header",
        "required": true,
        "schema": { "$ref": "#/components/schemas/Uuid" },
        "description": "Mock auth: user id returned by /auth/login."
      },
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "$ref": "#/components/schemas/Uuid" }
      },
      "ProductIdPath": {
        "name": "productId",
        "in": "path",
        "required": true,
        "schema": { "$ref": "#/components/schemas/Uuid" }
      },
      "Page": {
        "name": "page",
        "in": "query",
        "required": false,
        "schema": { "type": "integer", "minimum": 1, "default": 1 }
      },
      "PageSize": {
        "name": "pageSize",
        "in": "query",
        "required": false,
        "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
      }
    },
    "schemas": {
      "Uuid": {
        "type": "string",
        "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
      },
      "Money": { "type": "string", "pattern": "^\\d+\\.\\d{2}$" },
      "Timestamp": {
        "type": "string",
        "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(\\.\\d+)?Z$"
      },
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": ["code", "message", "details", "requestId"],
        "properties": {
          "code": { "type": "string", "pattern": "^[A-Z_]+$" },
          "message": { "type": "string" },
          "details": {},
          "requestId": { "type": "string" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "minLength": 3, "maxLength": 255 },
          "password": { "type": "string", "minLength": 1, "maxLength": 128 }
        }
      },
      "LoginResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["userId", "token", "role"],
        "properties": {
          "userId": { "$ref": "#/components/schemas/Uuid" },
          "token": { "type": "string" },
          "role": { "enum": ["customer", "admin"] }
        }
      },
      "Product": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "sku", "name", "category", "price", "isActive"],
        "properties": {
          "id": { "$ref": "#/components/schemas/Uuid" },
          "sku": { "type": "string" },
          "name": { "type": "string" },
          "category": { "type": "string" },
          "price": { "$ref": "#/components/schemas/Money" },
          "isActive": { "type": "boolean" }
        }
      },
      "ProductPage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["items", "page", "pageSize", "total"],
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/Product" } },
          "page": { "type": "integer", "minimum": 1 },
          "pageSize": { "type": "integer", "minimum": 1 },
          "total": { "type": "integer", "minimum": 0 }
        }
      },
      "Inventory": {
        "type": "object",
        "additionalProperties": false,
        "required": ["productId", "availableQuantity", "reservedQuantity"],
        "properties": {
          "productId": { "$ref": "#/components/schemas/Uuid" },
          "availableQuantity": { "type": "integer", "minimum": 0 },
          "reservedQuantity": { "type": "integer", "minimum": 0 }
        }
      },
      "CartItemRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["productId", "quantity"],
        "properties": {
          "productId": { "$ref": "#/components/schemas/Uuid" },
          "quantity": { "type": "integer", "minimum": 1, "maximum": 100 }
        }
      },
      "Cart": {
        "type": "object",
        "additionalProperties": false,
        "required": ["userId", "items"],
        "properties": {
          "userId": { "$ref": "#/components/schemas/Uuid" },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": false,
              "required": ["productId", "quantity"],
              "properties": {
                "productId": { "$ref": "#/components/schemas/Uuid" },
                "quantity": { "type": "integer", "minimum": 1 }
              }
            }
          }
        }
      },
      "CheckoutRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["userId", "items", "paymentMode"],
        "properties": {
          "userId": { "$ref": "#/components/schemas/Uuid" },
          "items": {
            "type": "array",
            "minItems": 1,
            "maxItems": 50,
            "items": {
              "type": "object",
              "additionalProperties": false,
              "required": ["productId", "quantity"],
              "properties": {
                "productId": { "$ref": "#/components/schemas/Uuid" },
                "quantity": { "type": "integer", "minimum": 1, "maximum": 100 }
              }
            }
          },
          "paymentMode": { "enum": ["MOCK_SUCCESS", "MOCK_FAIL", "MOCK_TIMEOUT"] }
        }
      },
      "OrderItem": {
        "type": "object",
        "additionalProperties": false,
        "required": ["productId", "quantity", "unitPrice"],
        "properties": {
          "productId": { "$ref": "#/components/schemas/Uuid" },
          "quantity": { "type": "integer", "minimum": 1 },
          "unitPrice": { "$ref": "#/components/schemas/Money" }
        }
      },
      "Order": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "userId",
          "status",
          "paymentStatus",
          "totalAmount",
          "items",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": { "$ref": "#/components/schemas/Uuid" },
          "userId": { "$ref": "#/components/schemas/Uuid" },
          "status": { "enum": ["PENDING", "CONFIRMED", "CANCELLED", "FAILED"] },
          "paymentStatus": { "enum": ["NOT_REQUIRED", "PENDING", "PAID", "FAILED", "REFUNDED"] },
          "totalAmount": { "$ref": "#/components/schemas/Money" },
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/OrderItem" } },
          "createdAt": { "$ref": "#/components/schemas/Timestamp" },
          "updatedAt": { "$ref": "#/components/schemas/Timestamp" }
        }
      },
      "OrderSummary": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "status", "paymentStatus", "totalAmount", "createdAt"],
        "properties": {
          "id": { "$ref": "#/components/schemas/Uuid" },
          "status": { "enum": ["PENDING", "CONFIRMED", "CANCELLED", "FAILED"] },
          "paymentStatus": { "enum": ["NOT_REQUIRED", "PENDING", "PAID", "FAILED", "REFUNDED"] },
          "totalAmount": { "$ref": "#/components/schemas/Money" },
          "createdAt": { "$ref": "#/components/schemas/Timestamp" }
        }
      },
      "OrderHistory": {
        "type": "object",
        "additionalProperties": false,
        "required": ["items", "page", "pageSize", "total"],
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/OrderSummary" } },
          "page": { "type": "integer", "minimum": 1 },
          "pageSize": { "type": "integer", "minimum": 1 },
          "total": { "type": "integer", "minimum": 0 }
        }
      },
      "PaymentRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["orderId", "amount", "mode"],
        "properties": {
          "orderId": { "$ref": "#/components/schemas/Uuid" },
          "amount": { "$ref": "#/components/schemas/Money" },
          "mode": { "enum": ["MOCK_SUCCESS", "MOCK_FAIL", "MOCK_TIMEOUT"] }
        }
      },
      "PaymentResult": {
        "type": "object",
        "additionalProperties": false,
        "required": ["paymentId", "orderId", "status", "duplicate"],
        "properties": {
          "paymentId": { "$ref": "#/components/schemas/Uuid" },
          "orderId": { "$ref": "#/components/schemas/Uuid" },
          "status": { "enum": ["PAID", "FAILED"] },
          "duplicate": {
            "type": "boolean",
            "description": "True when this orderId was already charged; the original result is returned and no second charge happens (INV-04)."
          }
        }
      },
      "Health": {
        "type": "object",
        "additionalProperties": false,
        "required": ["status", "service"],
        "properties": { "status": { "const": "ok" }, "service": { "type": "string" } }
      },
      "Ready": {
        "type": "object",
        "additionalProperties": false,
        "required": ["status", "checks"],
        "properties": {
          "status": { "enum": ["ready", "not-ready"] },
          "checks": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": false,
              "required": ["name", "ok"],
              "properties": { "name": { "type": "string" }, "ok": { "type": "boolean" } }
            }
          }
        }
      }
    },
    "responses": {
      "Error": {
        "description": "Standard error envelope",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } }
        }
      }
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "operationId": "login",
        "parameters": [{ "$ref": "#/components/parameters/RequestId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "Mock login",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/LoginResponse" } }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/products": {
      "get": {
        "operationId": "listProducts",
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PageSize" },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "maxLength": 80 }
          }
        ],
        "responses": {
          "200": {
            "description": "Active products",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProductPage" } }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/products/search": {
      "get": {
        "operationId": "searchProducts",
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "minLength": 1, "maxLength": 100 }
          },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PageSize" }
        ],
        "responses": {
          "200": {
            "description": "Matching active products",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProductPage" } }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "operationId": "getProduct",
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          { "$ref": "#/components/parameters/IdPath" }
        ],
        "responses": {
          "200": {
            "description": "Product",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Product" } }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/inventory/{productId}": {
      "get": {
        "operationId": "getInventory",
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          { "$ref": "#/components/parameters/ProductIdPath" }
        ],
        "responses": {
          "200": {
            "description": "Stock",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Inventory" } }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/cart/items": {
      "post": {
        "operationId": "upsertCartItem",
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          { "$ref": "#/components/parameters/UserIdHeader" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CartItemRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "Updated cart",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Cart" } } }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "409": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/cart": {
      "get": {
        "operationId": "getCart",
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          { "$ref": "#/components/parameters/UserIdHeader" }
        ],
        "responses": {
          "200": {
            "description": "Cart",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Cart" } } }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/cart/items/{productId}": {
      "delete": {
        "operationId": "removeCartItem",
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          { "$ref": "#/components/parameters/UserIdHeader" },
          { "$ref": "#/components/parameters/ProductIdPath" }
        ],
        "responses": {
          "200": {
            "description": "Updated cart",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Cart" } } }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/orders": {
      "post": {
        "operationId": "checkout",
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CheckoutRequest" } }
          }
        },
        "responses": {
          "201": {
            "description": "Order completed synchronously (monolith/REST). Replaying the same Idempotency-Key returns the same order.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Order" } }
            }
          },
          "202": {
            "description": "Order accepted (event-driven); status PENDING until the saga completes.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Order" } }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "402": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "409": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Error" },
          "504": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "operationId": "getOrder",
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          { "$ref": "#/components/parameters/IdPath" }
        ],
        "responses": {
          "200": {
            "description": "Order",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Order" } }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/orders/{id}/cancel": {
      "post": {
        "operationId": "cancelOrder",
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          { "$ref": "#/components/parameters/IdPath" }
        ],
        "responses": {
          "200": {
            "description": "Cancelled order; stock released (INV-05)",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Order" } }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "409": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/users/{id}/orders": {
      "get": {
        "operationId": "listUserOrders",
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          { "$ref": "#/components/parameters/IdPath" },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PageSize" }
        ],
        "responses": {
          "200": {
            "description": "Order history, newest first",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/OrderHistory" } }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/payments/mock": {
      "post": {
        "operationId": "mockPayment",
        "parameters": [{ "$ref": "#/components/parameters/RequestId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/PaymentRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "Payment result (idempotent per orderId)",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/PaymentResult" } }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "504": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/health": {
      "get": {
        "operationId": "health",
        "responses": {
          "200": {
            "description": "Liveness",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Health" } }
            }
          }
        }
      }
    },
    "/ready": {
      "get": {
        "operationId": "ready",
        "responses": {
          "200": {
            "description": "Ready",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Ready" } }
            }
          },
          "503": {
            "description": "Not ready",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Ready" } }
            }
          }
        }
      }
    }
  }
}